Sample Text File Instructions
This is an example txt data file that the program can take in. At the very top the first line of the file must contain the amount of terms in the file, then each next line shall contain first the number of google searches for the term and then the entire term with a maximum of six words. Commas are suggested between words but not required they just improve the look of outputs in the end.